+++ src/createModule.js
@@ -11,21 +11,20 @@
 |---------------------------------------------------
 */
 
-const glob     = require('glob');
-const chalk    = require('chalk');
+const glob = require('glob');
+const chalk = require('chalk');
 const {
   diff,
   addedDiff,
   deletedDiff,
   detailedDiff,
   updatedDiff
-}              = require('deep-object-diff');
-const path     = require('path');
-const memFs    = require('mem-fs');
-const editor   = require('mem-fs-editor');
-const findUp   = require('find-up');
-const _        = require('lodash');
-const inquirer = require('inquirer');
+} = require('deep-object-diff');
+const path = require('path');
+const memFs = require('mem-fs');
+const editor = require('mem-fs-editor');
+const findUp = require('find-up');
+const _ = require('lodash');
 
 const store = memFs.create();
 const fs = editor.create(store);
@@ -44,16 +43,16 @@ const createModule = ({
   flat = false
 }) => {
   let basePath = findUp.sync(['.ferdirc.js', '.ferdirc']);
-
   if (!basePath) {
-    console.error('Please create a config file named .ferdirc.js or .ferdirc in your project root');
+    console.error(
+      'Please create a config file named .ferdirc.js or .ferdirc in your project root',
+    );
     process.exit();
   }
 
   basePath = path.dirname(basePath);
-  const {
-    files
-  } = config;
+
+  const { files } = config;
 
   // function to copy a template file to a defined directory
   /**
@@ -61,10 +60,10 @@ const createModule = ({
    * @param fileExtension
    * @returns {*}
    */
-  const copyTpl = async ({
-    fileExtension
-  }) => {
-    const templatePath = path.resolve(`${basePath}/${config.paths.templateBase}`);
+  const copyTpl = ({ fileExtension }) => {
+    const templatePath = path.resolve(
+      `${basePath}/${config.paths.templateBase}`
+    );
 
     let globRegex = `?(*${kind}-*)`;
 
@@ -83,15 +82,22 @@ const createModule = ({
     if (pathOptions) destinationPath += pathOptions.path;
 
     const splitName = flat ? name.split('/') : name;
-    let fileName = '';
+
+    let fileName = ''
 
     if (flat) {
       fileName = splitName.pop();
     }
 
-    destinationPath = !flat ? path.join(destinationPath, name) : path.join(destinationPath, splitName.join('/'));
+    destinationPath = !flat ?
+      path.join(destinationPath, name) :
+      path.join(destinationPath, splitName.join('/'));
     // console.log(files[kind]);
-    let filename = files[kind].name ? files[kind].name : `${path.basename(name)}${files[kind].postfix ? `${files[kind].postfix}` : ''}`;
+    let filename = files[kind].name ?
+      files[kind].name :
+      `${path.basename(name)}${
+        files[kind].postfix ? `${files[kind].postfix}` : ''
+      }`;
 
     filename = `${filename}.${fileExtension}`;
     if (fileExtension.match(/scss/g) && !files[kind].name) {
@@ -116,23 +122,27 @@ const createModule = ({
     // }, null, 2));
 
     try {
-      if (fs.exists(destinationPath)) {
-        await inquirer.prompt([
-          { name: 'overwriteFile', type: 'confirm', message: 'You really want to overwrite the file?' }
-        ]);
-      }
-      // fs.copyTpl(templateFile[0], `${destinationPath.replace('//', '/')}`, moduleData);
-      console.log(chalk `\n{green File ${destinationPath.replace('//', '/')} was created}`);
+      fs.copyTpl(
+        templateFile[0],
+        `${destinationPath.replace('//', '/')}`,
+        moduleData
+      );
+      console.log(
+        chalk`\n{green File ${destinationPath.replace('//', '/')} was created}`
+      );
     } catch (error) {
       console.error(`No Template File found for ${kind}`, `\n${error}`);
     }
+
     return fs.commit(done => done);
   };
 
   if (typeof extension === Array) {
-    extension.forEach(extension => copyTpl({
-      fileExtension: extension
-    }));
+    extension.forEach(extension =>
+      copyTpl({
+        fileExtension: extension
+      })
+    );
   } else {
     copyTpl({
       fileExtension: extension
@@ -146,29 +156,22 @@ const createModule = ({
  * @param options
  * @param config
  */
-const moduleCreation = ({
-  options,
-  config
-}) => {
-  const {
-    files,
-    paths,
-    defaults
-  } = config;
-  const {
-    pathOptions
-  } = paths;
+const moduleCreation = ({ options, config }) => {
+  const { files, paths, defaults } = config;
+  const { pathOptions } = paths;
   let trueOptions = {};
 
-  const filteredOptions = Object.keys(options).filter(option => options[option]);
+  const filteredOptions = Object.keys(options).filter(
+    option => options[option]
+  );
 
-  const noDefaults = filteredOptions.some(item => Object.keys(defaults).indexOf(item) >= 0);
+  const noDefaults = filteredOptions.some(
+    item => Object.keys(defaults).indexOf(item) >= 0
+  );
   if (noDefaults) {
     trueOptions = options;
   } else {
-    trueOptions = { ...options,
-      ...defaults
-    };
+    trueOptions = { ...options, ...defaults };
   }
 
   Object.keys(files).forEach(file => {
@@ -181,10 +184,10 @@ const moduleCreation = ({
           path: files[file].path
         };
       } else {
-        Object.keys(pathOptions).forEach(path => {
+        Object.keys(pathOptions).forEach(pathKey => {
           if (options[path]) {
             destinationPathOption = {
-              key: path,
+              key: pathKey,
               path: pathOptions[path]
             };
           }
@@ -195,7 +198,9 @@ const moduleCreation = ({
 
       if (typeof options._[0] !== 'string') {
         console.error(`OPTIONS: ${JSON.stringify(trueOptions, null, 2)}`);
-        console.error(chalk `{red First argument must always be the name of the module}`);
+        console.error(
+          chalk`{red First argument must always be the name of the module}`
+        );
         return process.exit();
       }
 
@@ -208,7 +213,7 @@ const moduleCreation = ({
           pathOptions: destinationPathOption,
           flat: !!options.flat
         });
-      })
+      });
     }
   });
 };