+++ src/searchConfig.js
@@ -1,4 +1,4 @@
-const dir = require('node-dir');
+const dir = require("node-dir");
 
 module.exports = function searchFile(currentDir) {
   return new Promise((resolve, reject) => {
@@ -6,19 +6,19 @@ module.exports = function searchFile(currentDir) {
       currentDir,
       {
         match: /(config|package).json/g,
-        recursive: false,
+        recursive: false
       },
       function(err, content, filename, next) {
         if (err) throw err;
-        console.log('searching file');
-        if (!content.includes('moduleCreator')) {
-          console.log('nothing found');
-          reject('Oh no nothign found');
+        console.log("searching file");
+        if (!content.includes("moduleCreator")) {
+          console.log("nothing found");
+          reject("Oh no nothign found");
         }
-        console.log('found something');
+        console.log("found something");
         resolve(filename);
         next();
-      },
+      }
     );
   });
 };